Add experimentng pixel nerf and restructure the files
