+++ model/models.py
@@ -6,76 +6,6 @@ import imageio
 import json
 # from classification_models.keras import Classifiers
 
-
-# Misc utils
-
-def img2mse(x, y): return tf.reduce_mean(tf.square(x - y))
-
-
-def mse2psnr(x): return -10.*tf.math.log(x)/tf.math.log(10.)
-
-
-def to8b(x): return (255*np.clip(x, 0, 1)).astype(np.uint8)
-
-
-# Positional encoding
-
-class Embedder:
-
-    def __init__(self, **kwargs):
-
-        self.kwargs = kwargs
-        self.create_embedding_fn()
-
-    def create_embedding_fn(self):
-
-        embed_fns = []
-        d = self.kwargs['input_dims']
-        out_dim = 0
-        if self.kwargs['include_input']:
-            embed_fns.append(lambda x: x)
-            out_dim += d
-
-        max_freq = self.kwargs['max_freq_log2']
-        N_freqs = self.kwargs['num_freqs']
-
-        if self.kwargs['log_sampling']:
-            freq_bands = 2.**tf.linspace(0., max_freq, N_freqs)
-        else:
-            freq_bands = tf.linspace(2.**0., 2.**max_freq, N_freqs)
-
-        for freq in freq_bands:
-            for p_fn in self.kwargs['periodic_fns']:
-                embed_fns.append(lambda x, p_fn=p_fn,
-                                 freq=freq: p_fn(x * freq))
-                out_dim += d
-
-        self.embed_fns = embed_fns
-        self.out_dim = out_dim
-
-    def embed(self, inputs):
-        return tf.concat([fn(inputs) for fn in self.embed_fns], -1)
-
-
-def get_embedder(multires, i=0):
-
-    if i == -1:
-        return tf.identity, 3
-
-    embed_kwargs = {
-        'include_input': True,
-        'input_dims': 3,
-        'max_freq_log2': multires-1,
-        'num_freqs': multires,
-        'log_sampling': True,
-        'periodic_fns': [tf.math.sin, tf.math.cos],
-    }
-
-    embedder_obj = Embedder(**embed_kwargs)
-    def embed(x, eo=embedder_obj): return eo.embed(x)
-    return embed, embedder_obj.out_dim
-
-
 # Model architecture
 
 def init_nerf_model(D=8, W=256, input_ch=3, input_ch_views=3, output_ch=4, skips=[4], use_viewdirs=False):
@@ -266,7 +196,7 @@ def init_nerf_r_models(D=8, W=256, D_rotation=2, input_ch_image=(400, 400, 3), i
 
     return [model_encoder, model_decoder]
 
-def init_pixel_nerf_encoder(input_ch_image=(200, 200, 3)):
+def init_pixel_nerf_encoder(input_ch_image=(200, 200, 3), feature_depth=256):
 
     
     '''
@@ -295,7 +225,8 @@ def init_pixel_nerf_encoder(input_ch_image=(200, 200, 3)):
     pretrained_model.trainable = True
 
     # get features from middle layers
-    mid_out_layers = ['pool1_pool','conv2_block3_out', 'conv3_block4_out']#, 'conv4_block6_out', 'conv5_block3_out']
+    mid_out_layers = ['conv2_block3_out', 'conv3_block4_out', 'conv4_block6_out']
+    # mid_out_layers = ['pool1_pool','conv2_block3_out', 'conv3_block4_out', 'conv4_block6_out', 'conv5_block3_out']
     # mid_out_layers = ['stage1_unit2_conv2', 'stage2_unit2_conv2', 'stage3_unit2_conv2', 'stage4_unit2_conv2']
 
     features = []    
@@ -316,13 +247,13 @@ def init_pixel_nerf_encoder(input_ch_image=(200, 200, 3)):
     # pass feature for each pixel to a MLP to shrink size
     features_original_shape = features.shape
     features = tf.reshape(features, [-1, features_original_shape[-1]])
-    features = dense(64)(features)
+    features = dense(feature_depth)(features)
     
     model_encoder = tf.keras.Model(inputs=inputs_encoder, outputs=features, name='encoder')
 
     return model_encoder
 
-def init_pixel_nerf_decoder(D=8, W=256, input_ch_pose=(3,4), input_ch_coord=3, input_ch_views=3, output_ch=4, skips=[2,4,6], feature_depth=128):
+def init_pixel_nerf_decoder(D=8, W=256, input_ch_coord=3, input_ch_views=3, output_ch=4, skips=[2,4,6], feature_depth=256, mode='add'):
     '''
     Init decoder architecture from pixelNerf paper
     Uses skip connections for several times
@@ -341,123 +272,43 @@ def init_pixel_nerf_decoder(D=8, W=256, input_ch_pose=(3,4), input_ch_coord=3, i
     # -----------------------------------------------
     # Decoder MLP for predicting rbg and density
     inputs = tf.keras.Input(shape=(input_ch_coord + input_ch_views + feature_depth,))
+
+    # note that this can be further simplified
     inputs_pts, inputs_views, features = tf.split(inputs, [input_ch_coord, input_ch_views, feature_depth], -1)
     inputs_pts.set_shape([None, input_ch_coord])
     inputs_views.set_shape([None, input_ch_views])
     features.set_shape([None, feature_depth])
-
-    # feed features to an MLP first
-    # not needed -- done in encoder
-    # features = dense(128)(features)
-
     # concate feature vector with input coordinates
     decoder_inputs = tf.concat([inputs_pts,inputs_views, features], -1)
+
+    if mode == 'add':
+        # if using addition skip, need to ensure the shapes are the same
+        decoder_inputs = dense(W)(decoder_inputs)
+
     decoder_inputs_len = decoder_inputs.shape[-1]
     outputs = decoder_inputs
 
+
     for i in range(D):
-        if i+1 in skips:
-            # next layer has skip connection, need to shrink the size of this ouput
-            outputs = dense(W-decoder_inputs_len)(outputs)
-        else:
+        if mode == 'add':
             outputs = dense(W)(outputs)
-        
-        if i in skips:
-            outputs = tf.concat([outputs,decoder_inputs], -1)
+            if i in skips:
+                outputs = (outputs + decoder_inputs)/2 #? divide by 2 actually works!
+
+        else:
+            # use concate skip
+            if i+1 in skips:
+                # next layer has skip connection, need to shrink the size of this ouput
+                outputs = dense(W-decoder_inputs_len)(outputs)
+            else:
+                outputs = dense(W)(outputs)
+            
+            if i in skips:
+                outputs = tf.concat([outputs,decoder_inputs], -1)
+
+    outputs = dense(output_ch, act=None)(outputs)
 
     model_decoder = tf.keras.Model(inputs=inputs, outputs=outputs)
 
     return model_decoder
 
-
-# Ray helpers
-
-def get_rays(H, W, focal, c2w):
-    """Get ray origins, directions from a pinhole camera."""
-    i, j = tf.meshgrid(tf.range(W, dtype=tf.float32),
-                       tf.range(H, dtype=tf.float32), indexing='xy')
-    dirs = tf.stack([(i-W*.5)/focal, -(j-H*.5)/focal, -tf.ones_like(i)], -1)
-    rays_d = tf.reduce_sum(dirs[..., np.newaxis, :] * c2w[:3, :3], -1)
-    rays_o = tf.broadcast_to(c2w[:3, -1], tf.shape(rays_d))
-    return rays_o, rays_d
-
-
-def get_rays_np(H, W, focal, c2w):
-    """Get ray origins, directions from a pinhole camera."""
-    i, j = np.meshgrid(np.arange(W, dtype=np.float32),
-                       np.arange(H, dtype=np.float32), indexing='xy')
-    dirs = np.stack([(i-W*.5)/focal, -(j-H*.5)/focal, -np.ones_like(i)], -1)
-    rays_d = np.sum(dirs[..., np.newaxis, :] * c2w[:3, :3], -1)
-    rays_o = np.broadcast_to(c2w[:3, -1], np.shape(rays_d))
-    return rays_o, rays_d
-
-
-def ndc_rays(H, W, focal, near, rays_o, rays_d):
-    """Normalized device coordinate rays.
-
-    Space such that the canvas is a cube with sides [-1, 1] in each axis.
-
-    Args:
-      H: int. Height in pixels.
-      W: int. Width in pixels.
-      focal: float. Focal length of pinhole camera.
-      near: float or array of shape[batch_size]. Near depth bound for the scene.
-      rays_o: array of shape [batch_size, 3]. Camera origin.
-      rays_d: array of shape [batch_size, 3]. Ray direction.
-
-    Returns:
-      rays_o: array of shape [batch_size, 3]. Camera origin in NDC.
-      rays_d: array of shape [batch_size, 3]. Ray direction in NDC.
-    """
-    # Shift ray origins to near plane
-    t = -(near + rays_o[..., 2]) / rays_d[..., 2]
-    rays_o = rays_o + t[..., None] * rays_d
-
-    # Projection
-    o0 = -1./(W/(2.*focal)) * rays_o[..., 0] / rays_o[..., 2]
-    o1 = -1./(H/(2.*focal)) * rays_o[..., 1] / rays_o[..., 2]
-    o2 = 1. + 2. * near / rays_o[..., 2]
-
-    d0 = -1./(W/(2.*focal)) * \
-        (rays_d[..., 0]/rays_d[..., 2] - rays_o[..., 0]/rays_o[..., 2])
-    d1 = -1./(H/(2.*focal)) * \
-        (rays_d[..., 1]/rays_d[..., 2] - rays_o[..., 1]/rays_o[..., 2])
-    d2 = -2. * near / rays_o[..., 2]
-
-    rays_o = tf.stack([o0, o1, o2], -1)
-    rays_d = tf.stack([d0, d1, d2], -1)
-
-    return rays_o, rays_d
-
-
-# Hierarchical sampling helper
-
-def sample_pdf(bins, weights, N_samples, det=False):
-
-    # Get pdf
-    weights += 1e-5  # prevent nans
-    pdf = weights / tf.reduce_sum(weights, -1, keepdims=True)
-    cdf = tf.cumsum(pdf, -1)
-    cdf = tf.concat([tf.zeros_like(cdf[..., :1]), cdf], -1)
-
-    # Take uniform samples
-    if det:
-        u = tf.linspace(0., 1., N_samples)
-        u = tf.broadcast_to(u, list(cdf.shape[:-1]) + [N_samples])
-    else:
-        u = tf.random.uniform(list(cdf.shape[:-1]) + [N_samples])
-
-    # Invert CDF
-    inds = tf.searchsorted(cdf, u, side='right')
-    below = tf.maximum(0, inds-1)
-    above = tf.minimum(cdf.shape[-1]-1, inds)
-    inds_g = tf.stack([below, above], -1)
-    cdf_g = tf.gather(cdf, inds_g, axis=-1, batch_dims=len(inds_g.shape)-2)
-    bins_g = tf.gather(bins, inds_g, axis=-1, batch_dims=len(inds_g.shape)-2)
-
-    denom = (cdf_g[..., 1]-cdf_g[..., 0])
-    denom = tf.where(denom < 1e-5, tf.ones_like(denom), denom)
-    t = (u-cdf_g[..., 0])/denom
-    samples = bins_g[..., 0] + t * (bins_g[..., 1]-bins_g[..., 0])
-
-    return samples
