+++ test_run_init_nerf_r.py
@@ -1,4 +1,4 @@
-from run_nerf_helpers import *
+from model.models import *
 import numpy as np
