+++ utils/run_nerf_helpers.py
@@ -1,0 +1,109 @@
+import os
+import sys
+import tensorflow as tf
+import numpy as np
+import imageio
+import json
+
+
+# Misc utils
+
+def img2mse(x, y): return tf.reduce_mean(tf.square(x - y))
+
+
+def mse2psnr(x): return -10.*tf.math.log(x)/tf.math.log(10.)
+
+
+def to8b(x): return (255*np.clip(x, 0, 1)).astype(np.uint8)
+
+# Ray helpers
+
+def get_rays(H, W, focal, c2w):
+    """Get ray origins, directions from a pinhole camera."""
+    i, j = tf.meshgrid(tf.range(W, dtype=tf.float32),
+                       tf.range(H, dtype=tf.float32), indexing='xy')
+    dirs = tf.stack([(i-W*.5)/focal, -(j-H*.5)/focal, -tf.ones_like(i)], -1)
+    rays_d = tf.reduce_sum(dirs[..., np.newaxis, :] * c2w[:3, :3], -1)
+    rays_o = tf.broadcast_to(c2w[:3, -1], tf.shape(rays_d))
+    return rays_o, rays_d
+
+
+def get_rays_np(H, W, focal, c2w):
+    """Get ray origins, directions from a pinhole camera."""
+    i, j = np.meshgrid(np.arange(W, dtype=np.float32),
+                       np.arange(H, dtype=np.float32), indexing='xy')
+    dirs = np.stack([(i-W*.5)/focal, -(j-H*.5)/focal, -np.ones_like(i)], -1)
+    rays_d = np.sum(dirs[..., np.newaxis, :] * c2w[:3, :3], -1)
+    rays_o = np.broadcast_to(c2w[:3, -1], np.shape(rays_d))
+    return rays_o, rays_d
+
+
+def ndc_rays(H, W, focal, near, rays_o, rays_d):
+    """Normalized device coordinate rays.
+
+    Space such that the canvas is a cube with sides [-1, 1] in each axis.
+
+    Args:
+      H: int. Height in pixels.
+      W: int. Width in pixels.
+      focal: float. Focal length of pinhole camera.
+      near: float or array of shape[batch_size]. Near depth bound for the scene.
+      rays_o: array of shape [batch_size, 3]. Camera origin.
+      rays_d: array of shape [batch_size, 3]. Ray direction.
+
+    Returns:
+      rays_o: array of shape [batch_size, 3]. Camera origin in NDC.
+      rays_d: array of shape [batch_size, 3]. Ray direction in NDC.
+    """
+    # Shift ray origins to near plane
+    t = -(near + rays_o[..., 2]) / rays_d[..., 2]
+    rays_o = rays_o + t[..., None] * rays_d
+
+    # Projection
+    o0 = -1./(W/(2.*focal)) * rays_o[..., 0] / rays_o[..., 2]
+    o1 = -1./(H/(2.*focal)) * rays_o[..., 1] / rays_o[..., 2]
+    o2 = 1. + 2. * near / rays_o[..., 2]
+
+    d0 = -1./(W/(2.*focal)) * \
+        (rays_d[..., 0]/rays_d[..., 2] - rays_o[..., 0]/rays_o[..., 2])
+    d1 = -1./(H/(2.*focal)) * \
+        (rays_d[..., 1]/rays_d[..., 2] - rays_o[..., 1]/rays_o[..., 2])
+    d2 = -2. * near / rays_o[..., 2]
+
+    rays_o = tf.stack([o0, o1, o2], -1)
+    rays_d = tf.stack([d0, d1, d2], -1)
+
+    return rays_o, rays_d
+
+
+# Hierarchical sampling helper
+
+def sample_pdf(bins, weights, N_samples, det=False):
+
+    # Get pdf
+    weights += 1e-5  # prevent nans
+    pdf = weights / tf.reduce_sum(weights, -1, keepdims=True)
+    cdf = tf.cumsum(pdf, -1)
+    cdf = tf.concat([tf.zeros_like(cdf[..., :1]), cdf], -1)
+
+    # Take uniform samples
+    if det:
+        u = tf.linspace(0., 1., N_samples)
+        u = tf.broadcast_to(u, list(cdf.shape[:-1]) + [N_samples])
+    else:
+        u = tf.random.uniform(list(cdf.shape[:-1]) + [N_samples])
+
+    # Invert CDF
+    inds = tf.searchsorted(cdf, u, side='right')
+    below = tf.maximum(0, inds-1)
+    above = tf.minimum(cdf.shape[-1]-1, inds)
+    inds_g = tf.stack([below, above], -1)
+    cdf_g = tf.gather(cdf, inds_g, axis=-1, batch_dims=len(inds_g.shape)-2)
+    bins_g = tf.gather(bins, inds_g, axis=-1, batch_dims=len(inds_g.shape)-2)
+
+    denom = (cdf_g[..., 1]-cdf_g[..., 0])
+    denom = tf.where(denom < 1e-5, tf.ones_like(denom), denom)
+    t = (u-cdf_g[..., 0])/denom
+    samples = bins_g[..., 0] + t * (bins_g[..., 1]-bins_g[..., 0])
+
+    return samples
